join nl together
